requirements/test.txt
# Test dependencies go here.
-r base.txt
coverage==3.6
django-discover-runner==0.4
django-debug-toolbar==1.0.1
django-crispy-forms==1.4.0
